# backend/docs/contract/API.yml
openapi: 3.0.3
info:
  title: 이걸주네? API
  version: 1.0.0
  description: |
    AI 맞춤형 선물 추천 서비스의 MVP REST API 명세입니다.
    일반 이메일·비밀번호 회원가입과 로그인을 사용합니다.
    카카오톡 원문은 분석 중에만 임시 처리하고 영구 저장하지 않습니다.
  contact:
    name: 이걸주네? 프로젝트 팀
    url: https://github.com/beshurl/igeoljune
  license:
    name: All rights reserved
    url: https://github.com/beshurl/igeoljune
servers:
  - url: /
    description: 배포 호스트와 로컬·Mock 호스트에 공통으로 적용하는 루트 URL
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/recipients/{recipientId}/preferences:
    parameters:
      - $ref: '#/components/parameters/RecipientId'
    post:
      tags: [Preferences]
      summary: 직접 입력 취향 등록
      description: sourceType은 서버에서 DIRECT로 설정합니다.
      operationId: createPreference
      x-uc-id: UC3
      x-req-id: REQ-F-003
      x-screen-id: SCR-GIFT-001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenceCreateRequest'
      responses:
        '201':
          description: 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredPreference'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [Preferences]
      summary: 취향 목록 조회
      operationId: listPreferences
      x-uc-id: UC3,UC6
      x-req-id: REQ-F-003,REQ-F-006
      x-screen-id: SCR-GIFT-001,SCR-KAKAO-002
      parameters:
        - name: preferenceType
          in: query
          schema:
            $ref: '#/components/schemas/PreferenceType'
        - name: sourceType
          in: query
          schema:
            $ref: '#/components/schemas/SourceType'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferenceListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/recipients/{recipientId}/preferences/bulk:
    parameters:
      - $ref: '#/components/parameters/RecipientId'
    post:
      tags: [Preferences]
      summary: 승인된 카카오 추출 취향 일괄 저장
      description: 카카오 분석 검토 화면에서 승인한 항목만 저장하며 sourceType은 서버가 KAKAO로 설정합니다.
      operationId: createPreferencesBulk
      x-uc-id: UC6
      x-req-id: REQ-F-006
      x-screen-id: SCR-KAKAO-002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenceBulkCreateRequest'
      responses:
        '201':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferenceListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/preferences/{preferenceId}:
    parameters:
      - $ref: '#/components/parameters/PreferenceId'
    patch:
      tags: [Preferences]
      summary: 취향 수정
      operationId: updatePreference
      x-uc-id: UC6
      x-req-id: REQ-F-006
      x-screen-id: SCR-KAKAO-002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenceUpdateRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredPreference'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Preferences]
      summary: 취향 삭제
      operationId: deletePreference
      x-uc-id: UC6
      x-req-id: REQ-F-006
      x-screen-id: SCR-KAKAO-002
      responses:
        '204':
          description: 삭제 성공
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RecipientId:
      name: recipientId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 12
    PreferenceId:
      name: preferenceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 31
...
  responses:
    BadRequest:
      description: 입력값 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: 입력값을 확인해 주세요.
            fieldErrors:
              - field: budgetMax
                reason: budgetMin 이상이어야 합니다.
...
    Forbidden:
      description: 다른 사용자의 리소스 접근
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: RESOURCE_FORBIDDEN
            message: 접근할 수 없는 리소스입니다.
            fieldErrors: []
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: RESOURCE_NOT_FOUND
            message: 요청한 리소스를 찾을 수 없습니다.
            fieldErrors: []
    Conflict:
      description: 이메일·중복 데이터·리소스 상태 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: RESOURCE_CONFLICT
            message: 현재 상태에서는 요청을 처리할 수 없습니다.
            fieldErrors: []
...
  schemas:
...
    PreferenceType:
      type: string
      enum: [INTEREST, PREFERRED_CATEGORY, PREFERRED_ATTRIBUTE, DISLIKED_CATEGORY, WISH_ITEM]
    SourceType:
      type: string
      enum: [DIRECT, KAKAO]
    PreferenceCreateRequest:
      type: object
      required: [preferenceType, preferenceValue]
      additionalProperties: false
      properties:
        preferenceType:
          $ref: '#/components/schemas/PreferenceType'
        preferenceValue:
          type: string
          minLength: 1
          maxLength: 255
      example:
        preferenceType: INTEREST
        preferenceValue: 러닝
    PreferenceUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        preferenceType:
          $ref: '#/components/schemas/PreferenceType'
        preferenceValue:
          type: string
          minLength: 1
          maxLength: 255
    PreferenceBulkCreateRequest:
      type: object
      required: [items]
      additionalProperties: false
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/PreferenceCreateRequest'
      example:
        items:
          - preferenceType: INTEREST
            preferenceValue: 홈카페
          - preferenceType: DISLIKED_CATEGORY
            preferenceValue: 향수
    StructuredPreference:
      type: object
      required: [preferenceId, recipientId, preferenceType, preferenceValue, sourceType, createdAt, updatedAt]
      properties:
        preferenceId:
          type: integer
          format: int64
        recipientId:
          type: integer
          format: int64
        preferenceType:
          $ref: '#/components/schemas/PreferenceType'
        preferenceValue:
          type: string
        sourceType:
          $ref: '#/components/schemas/SourceType'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        preferenceId: 31
        recipientId: 12
        preferenceType: INTEREST
        preferenceValue: 홈카페
        sourceType: KAKAO
        createdAt: '2026-09-03T14:31:00+09:00'
        updatedAt: '2026-09-03T14:31:00+09:00'
    PreferenceListResponse:
      type: object
      required: [items, totalCount]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StructuredPreference'
        totalCount:
          type: integer
          minimum: 0
      example:
        items:
          - preferenceId: 31
            recipientId: 12
            preferenceType: INTEREST
            preferenceValue: 홈카페
            sourceType: KAKAO
            createdAt: '2026-09-03T14:31:00+09:00'
            updatedAt: '2026-09-03T14:31:00+09:00'
        totalCount: 1
...
    FieldError:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    ErrorResponse:
      type: object
      required: [code, message, fieldErrors]
      properties:
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
      example:
        code: RESOURCE_NOT_FOUND
        message: 요청한 리소스를 찾을 수 없습니다.
        fieldErrors: []