# backend/src/main/java/com/skala/ikgeoljune/domain/SourceType.java
package com.skala.ikgeoljune.domain;

/** API 명세서 §10 sourceType */
public enum SourceType {
    /** 사용자 직접 입력 */
    DIRECT,
    /** 카카오톡 대화 분석 결과 */
    KAKAO
}